==> 3_Downstream_visualization.Rmd <==
# Downstream Visualizations 
> **Description**: This pipeline performs downstream visualizations, including volcano plot, heatmap, and box plots.


**Project Initialization**:
```{r ,warning = FALSE, message = FALSE}
#Sets the working directory and creates subfolders for organizing outputs.
mypath= "C:/Users/USER/Documents/Github/CRC_project/"
dir.create("output")
dir.create("plots")
dir.create("input")

```

```{r ,warning = FALSE, message = FALSE}
#Load libraries
library(tibble)
library(plyr)
library(dplyr)
library(tidyverse)
library(openxlsx)
library(cowplot)
library(ggplot2)
library(tibble)
library("RColorBrewer")
library("circlize")
library(ComplexHeatmap)
library(biomaRt)
library(dplyr)
library(plyr)
library(fields)
library(tidyr)
library(RColorBrewer)
library(viridis)
library(ggplot2)
library(ggrepel)


```


```{r ,warning = FALSE, message = FALSE}
#load data
data= read.csv(paste0(mypath,"input/data_for_downstream.csv")) 
data = data %>% column_to_rownames(colnames(data)[1]) %>% dplyr::select(-1)
de=   read.csv(paste0(mypath,"output/DE_sig.csv"))
de=de |> filter(abs(logFC) >= log2(3))
output_dir= paste0(mypath, "output/")

# set group distribution for samples
group_dist= gsub("_.*", "", colnames(data))
print(group_dist)
group_levels= unique(group_dist)

group_colors <- c("#fc8d62", "#66c2a5")
names(group_colors) <- group_levels


```

## **Volcano Plot**  
```{r ,warning = FALSE, message = FALSE}
plot_volcano = function(file_path=NULL, df=NULL, name=NULL, 
                        plot_dir = "plots", fc_threshold = 1.5) {
  
  if (!dir.exists(plot_dir)) dir.create(plot_dir, recursive = TRUE)
    # Read the DE result
  if(!is.null(file_path)){
    res <- read.csv(file_path)
    name <- gsub(".*_([[:alnum:] ]+)\\.csv$", "\\1", basename(file_path))
  }
  
  if(!is.null(df) & !is.null(name)){
    res= df
    name= name
  }
    res <- as.data.frame(res)
    colnames(res)[grepl("log",colnames(res) )]= "logFC"
    colnames(res)[grepl("adj|FDR",colnames(res) )]= "padj"
    res$padj[is.na(res$padj)]= res$pval[is.na(res$padj)]
    
    # Define direction
    res$Direction <- ifelse(res$padj <= 0.05 & res$logFC >= log2(fc_threshold), "Up",
                            ifelse(res$padj <= 0.05 & res$logFC <= -log2(fc_threshold), "Down",
                                                                             "Non-significant"))
    # Keep only labels for up or down expressed features
    res <- res %>% mutate(features = ifelse(Direction %in% c("Up", "Down"), X, ""))
    # Compute -log10(p-value)
    res$log10p <- -log10(res$padj)
    # Plot parameters
    xminma <- -3.5
    xmaxma <- 3.5
    
    # Generate the plot
    volcano <- ggplot(res, aes(x = logFC, y = log10p, color = Direction, label = features)) +
      geom_point(size = 1.2, alpha = 0.7) +
      geom_rug(alpha = 0.6) +
      scale_color_manual(values = c("Up" = "red2", "Down" = "darkslateblue", "Non-significant" = "grey66")) +
      xlab('log2 Fold Change') +
      ylab('-log10 adj.p-value') +
      #scale_x_continuous(limits = c(xminma, xmaxma)) +
      theme_bw() +
      theme(legend.title = element_blank()) +
      geom_vline(xintercept = c(-log2(fc_threshold), 0, log2(fc_threshold)), linetype = c("dotted", "solid", "dotted")) +
      geom_hline(yintercept = -log10(0.05), linetype = "dashed") +
      geom_text_repel(aes(label = features),
                      size = 2,
                      max.overlaps = 10,
                      segment.color = "grey50",
                      color = "black") +
      ggtitle(paste0("Volcano Plot of DE ", name))
    
    # Print and save
    print(volcano)
    ggsave(filename = paste0(plot_dir, "/Volcanoplot_", name, ".jpg"),
           plot = volcano,
           dpi = 600,
           width = 7,
           height = 4)
  }



```

```{r ,warning = FALSE, message = FALSE}
#Apply the function
plot_volcano(file_path= paste0(mypath,"output/DE_all.csv"))


```

## **Heatmap Plot**
```{r ,warning = FALSE, message = FALSE}
plot_heatmap= function(data= NULL, de= NULL, group_dist, group_colors ){ 
  
    #you can use for loop for multiple groups DE files
    data_sub= data[row.names(data) %in% de$X, ]
    #data= data %>% dplyr::slice(1:50)
    #row.names= gsub("\\%.*", "", row.names(data))
    heat_data=  t(scale(t(data_sub)))  # center row wise
    rownames(heat_data)=  substr(rownames(heat_data), 1, 25) 

    ta <- HeatmapAnnotation(
      Condition = group_dist,
      col = list(
        Condition = group_colors
        
      ),
      annotation_height = unit(10, "mm")
    )
    
    #palt=  colorRampPalette(c("blue4", "black", "#8B0000"))(256)
    #palt= colorRampPalette(viridis(n = 8))(200)
    palt= colorRamp2(c(-3, 0, 3), c("#053061", "#F7F5F4", "darkred"))
    
    heatmap  <- Heatmap(
      matrix = as.matrix(heat_data),
      name = "Normalized Intensity",
      col =palt ,#' ## ' #' ## #' ## matlab::jet.colors(200),
      show_row_names = TRUE,
      cluster_rows = TRUE,
      cluster_columns = TRUE,
      show_column_names = FALSE,
      row_names_gp = gpar(fontsize = 8),
      top_annotation  = ta,
      heatmap_legend_param = list(
        title= "Scaled\nIntensity",
        title_gp = gpar(fontsize = 10, fontface = "bold"),  
        labels_gp = gpar(fontsize = 10),
        legend_height = unit(4, "cm"),   
        legend_width = unit(.5, "cm")   
      )
    ) 
    print(heatmap)
    png("plots/heatmap.png",width = 4000, height = 5000, res = 600)
    draw(heatmap, annotation_legend_side = "right", heatmap_legend_side = "right") #"right"
    dev.off()
    
}


```

```{r ,warning = FALSE, message = FALSE}
#Apply the function
plot_heatmap(data, de, group_dist, group_colors)


```
## **Box Plots**
```{r ,warning = FALSE, message = FALSE}
# Loop over DE features
plot_box= function( data=NULL, de= NULL, group_dist){
   
    data= data[row.names(data) %in% de$X, ]
    data.t= t(data)
    de <- as.data.frame(de)
    colnames(de)[grepl("log",colnames(de) )]= "logFC"
    colnames(de)[grepl("adj|FDR",colnames(de) )]= "padj"
    de$padj[is.na(de$padj)]= de$pval[is.na(de$padj)]
    
for (m in colnames(data.t)) {
    # Subset data
    data.box = data.frame(
            value= data.t[, m],
            group= group_dist)
    
    FDR= de$padj[de$X== m]
    quartile.t= quantile(  data.box$value, 0.90)
    
    p= ggplot(data.box , aes(x = group , y = value , fill= group)) +
      geom_boxplot() +
      scale_fill_manual(values = group_colors )+
      labs(title = paste0("Box Plot of Intensity for ", m, " Metabolite"), x = "",
           y = "Intensity",
           fill= "Groups")+
      annotate("text", 
               x = 3.5, 
               y =  quartile.t, 
               label = paste0( "FDR= " , round(FDR, 2)) , 
               size = 3, fontface = "italic", hjust = 1) +
      theme_minimal()+
      theme(
        axis.text.x = element_text(angle = 0, hjust = .65, size = 7),
        axis.text.y = element_text(size = 7),
        plot.title = element_text(size = 9),
        axis.line = element_line(color = "black"),
        panel.grid = element_blank(),
        axis.ticks = element_blank(),
        panel.border = element_blank())
    
    print(p)
    ggsave(paste0("plots/boxplot_" , m, ".png"), p, width=5, height =4, dpi=600)
    
  }
}



```

```{r ,warning = FALSE, message = FALSE}
#Apply the function
plot_box(data, de[1:3,], group_dist )

```

```{r, echo=FALSE}
sessionInfo()
```
